socialmedia/requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
